+++ src/mkvtoolnix-gui/merge/source_file.h
@@ -69,6 +69,9 @@ public:
 protected:
   virtual QString deriveLanguageFromFileName();
   virtual QRegularExpression regexForDerivingLanguageFromFileName();
+
+public:
+  static QString defaultRegexForDerivingLanguageFromFileName();
 };
 
 }}}